puppet: prevent unnecessary restarts/reinstalls on windows

+++ deployments/puppet/manifests/win_repo.pp
@@ -3,34 +3,52 @@ class signalfx_agent::win_repo (
   $repo_base,
   $package_stage,
   $version,
-  $config_file_path,
   $installation_directory,
   $service_name,
 ) {
 
-  $url = "https://${repo_base}/windows/${package_stage}/zip/SignalFxAgent-${version}-win64.zip"
-  $zipfile_location = "${installation_directory}\\SignalFxAgent-${version}-win64.zip"
+  $zipfile_name = "SignalFxAgent-${version}-win64.zip"
+  $url = "https://${repo_base}/windows/${package_stage}/zip/${zipfile_name}"
+  $zipfile_location = "${installation_directory}\\${zipfile_name}"
+  $version_file = "${installation_directory}\\version.txt"
 
   file { $installation_directory:
     ensure => 'directory',
   }
 
-  -> exec { 'Stop SignalFx Agent':
-    command  => "Stop-Service -Name \'${service_name}\'",
-    onlyif   => "((Get-CimInstance -ClassName win32_service -Filter 'Name = \'${service_name}\'' | Select Name, State).Name)",
-    provider => 'powershell',
+  if find_file($version_file) {
+    $installed_version = file($version_file)
+  }
+  else {
+    $installed_version = ''
   }
 
-  -> archive { $zipfile_location:
-    source       => $url,
-    extract_path => $installation_directory,
-    group        => 'Administrator',
-    user         => 'Administrator',
-    extract      => true,
-  }
+  if $url != $installed_version {
+    exec { 'Stop SignalFx Agent':
+      command  => "Stop-Service -Name \'${service_name}\'",
+      onlyif   => "((Get-CimInstance -ClassName win32_service -Filter 'Name = \'${service_name}\'' | Select Name, State).Name)",
+      provider => 'powershell',
+    }
 
-  -> tidy { $installation_directory:
-    recurse => 1,
-    matches => ['*.zip'],
+    -> archive { $zipfile_name:
+      source       => $url,
+      path         => $zipfile_location,
+      extract_path => $installation_directory,
+      group        => 'Administrator',
+      user         => 'Administrator',
+      extract      => true,
+      cleanup      => true,
+      require      => File[$installation_directory],
+    }
+
+    -> file { $version_file:
+      content => $url,
+    }
+
+    # ensure zip file is always deleted
+    -> file { "Delete ${zipfile_location}":
+      ensure => 'absent',
+      path   => $zipfile_location,
+    }
   }
 }
